Create file for TCP-specific functions
